Refactoring
- Move item creating logic into action
- Move skill GetSkill() to SkillFactory
- Random elemental type for custom crafted equipment
