+++ Lib9c/Model/Item/ItemBase.cs
@@ -92,6 +92,8 @@ namespace Nekoyume.Model.Item
 
                 return _elementalType;
             }
+
+            set => _elementalType = value;
         }
 
         protected ItemBase(ItemSheet.Row data)
